requirements.txt
asgiref==3.6.0
boto3==1.26.38
botocore==1.29.38
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
